multi: address lint findings from CI for #270 branch
Mechanical lint cleanups surfaced on CI for the seal-time fee
handshake branch:

  - rounds/fsm_transitions.go: rename local `cap` (both declaration
    and references) to `passCap` to stop shadowing the Go
    predeclared builtin; drop redundant `x = x + 1` for `x++`;
    extract the test-escape-hatch NextState / Outbox values to
    local variables so the nested struct literal fits under 80
    cols; add blank lines before the seal-time drop-path
    `continue`s.
  - rounds/fsm_states.go: wrap newIntentCollectingState signature
    and trim the doc comment so both fit under 80 cols.
  - rounds/seal_time_fee_builder.go: retag package doc with
    "Package rounds " prefix so godoclint accepts it; nlreturn
    blank line before return inside locateChangeDesignation.
  - rounds/seal_time_fee_builder_test.go: wrap the long
    BoardingInputs struct literals and shorten the rapid
    balance-invariant assertion message.
  - rounds/outbox_messages_quote_test.go: wrap the require.True
    message inside the positional-order iteration loop.
  - rounds/quote_sent_state_test.go: fix dupword ("env env") in
    driveAllQuotesResolved doc; convert unchecked type assertions
    (`ns := tr.NextState.(*QuoteSentState)` etc.) to the
    two-result form required by forcetypeassert; rewrap
    require.Contains calls that crossed 80 cols.
  - rounds/fsm_transitions_test.go: trim comments that tripped the
    line-length rule.
  - server_rounds.go: drop the stale //nolint:funlen directive on
    registerQuoteRoutes (the helper is no longer long enough to
    trip funlen).
  - db/vtxo_store_test.go: drop the redundant uint32() conversion
    on testRound.CSVDelay (already uint32).

No behaviour change; unit tests pass.

+++ db/vtxo_store_test.go
@@ -125,7 +125,7 @@ func TestVTXOStoreLoadBatchExpiry(t *testing.T) {
 		require.NoError(t, err)
 		require.NotNil(t, loaded)
 		require.Equal(
-			t, uint32(confHeight)+uint32(testRound.CSVDelay),
+			t, uint32(confHeight)+testRound.CSVDelay,
 			loaded.BatchExpiry,
 		)
 	})
